requirements.txt
amqp==2.2.1
anyjson==0.3.3
asgi-redis==1.4.2
asgiref==1.1.2
asn1crypto==0.22.0
astroid==1.5.3
attrs==17.2.0
autobahn==17.6.2
Automat==0.6.0
awscli==1.11.136
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.3
boto3==1.4.4
botocore==1.6.3
celery==4.1.0
certifi==2017.7.27.1
cffi==1.10.0
channels==1.1.6
chardet==3.0.4
click==6.7
colorama==0.3.7
constantly==15.1.0
coreapi==2.3.0
coreschema==0.0.4
cryptography==2.0
daphne==1.3.0
decorator==4.0.10
defusedxml==0.5.0
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.11
django-allauth==0.32.0
django-appconf==1.0.2
django-celery-beat==1.0.1
django-compressor==2.1.1
django-cors-headers==1.3.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-extensions==1.7.8
django-filter==1.0.4
django-haystack==2.6.1
django-rest-swagger==2.1.2
djangorestframework==3.6.2
docutils==0.13.1
elasticsearch==2.3.0
enum34==1.1.6
flake8==3.4.1
futures==3.1.1
gunicorn==19.6.0
h2==3.0.1
hpack==3.0.0
hyperframe==5.1.0
hyperlink==17.3.0
idna==2.6
incremental==17.5.0
ipython==5.1.0
ipython-genutils==0.1.0
isort==4.2.15
itypes==1.1.0
Jinja2==2.9.6
jmespath==0.9.3
kombu==4.1.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
msgpack-python==0.4.8
mysqlclient==1.3.10
oauthlib==2.0.2
openapi-codec==1.3.1
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
priority==1.3.0
prompt-toolkit==1.0.9
psycopg2==2.7.1
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.5.0
Pygments==2.1.3
PyJWT==1.5.2
pylint==1.7.2
pyOpenSSL==17.2.0
python-dateutil==2.6.0
python-dotenv==0.6.4
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
rcssmin==1.0.6
redis==2.10.5
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
rsa==3.4.2
s3transfer==0.1.10
service-identity==17.0.0
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.2
static3==0.7.0
stripe==1.62.1
traitlets==4.3.1
Twisted==17.5.0
txaio==2.8.0
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.15
vine==1.1.3
wcwidth==0.1.7
whitenoise==3.2.2
wrapt==1.10.10
zope.interface==4.4.2
